20240124_LaConcordiaRed/requirements.txt
pandas~=2.2.0
numpy~=1.26.3
matplotlib~=3.8.2
mplsoccer~=1.2.2
pillow~=10.2.0